Implement User endpoints
Eager load users weather in the model for simplicity -- it's a has
one relationship and we're paginatig users, so the performance hit
should be negligible.

I've considered the repository pattern to show off coding chops, but
I didn't want to give the impression that I default to it in every
project like the community often does. In my experience it's overkill
most of the time.